+++ lib/theme/durations.dart
@@ -61,7 +61,6 @@ class ADurations {
   static const highlightScrollInitDelay = Duration(milliseconds: 800);
   static const motionPhotoAutoPlayDelay = Duration(milliseconds: 700);
   static const appInactiveReactionDelay = Duration(milliseconds: 300);
-  static const videoOverlayHideDelay = Duration(milliseconds: 500);
   static const videoProgressTimerInterval = Duration(milliseconds: 300);
   static const doubleBackTimerDelay = Duration(milliseconds: 1000);
   static const softKeyboardDisplayDelay = Duration(milliseconds: 300);